fix(rcv): Replace `float` with `Decimal` for `tasa_otro_impuesto`
- Updated `tasa_otro_impuesto` data type to `Decimal` for improved precision.
- Removed redundant validator for rounding `tasa_otro_impuesto`.
- Adjusted parsers and tests for the new data type.

Ref: https://app.shortcut.com/cordada/story/16788/

+++ src/cl_sii/rcv/data_models.py
@@ -9,6 +9,7 @@ from __future__ import annotations
 
 import logging
 from datetime import date, datetime
+from decimal import Decimal
 from typing import ClassVar, Optional
 
 import pydantic
@@ -546,7 +547,7 @@ class RcDetalleEntry(RcvDetalleEntry):
     Valor Otro Impuesto
     """
 
-    tasa_otro_impuesto: Optional[float]
+    tasa_otro_impuesto: Optional[Decimal]
     """
     Tasa Otro Impuesto
     """
